Fixes issue 2963, ref tcp_connecter.cpp:423

+++ src/tcp_connecter.cpp
@@ -435,6 +435,6 @@ void zmq::tcp_connecter_t::close ()
     const int rc = ::close (s);
     errno_assert (rc == 0);
 #endif
-    socket->event_closed (endpoint, (int) s);
+    socket->event_closed (endpoint, s);
     s = retired_fd;
 }
